Project structure adjustments
